+++ src/Eras.Application/Features/Variables/Queries/GetVariablesByPollUuidAndComponent/GetVariablesByPollUuidAndComponentQuery.cs
@@ -1,0 +1,10 @@
+﻿using Eras.Domain.Entities;
+using MediatR;
+
+namespace Eras.Application.Features.Variables.Queries.GetVariablesByPollUuidAndComponent
+{
+    public sealed record GetVariablesByPollUuidAndComponentQuery(
+        string pollUuid,
+        List<string> component
+    ) : IRequest<List<Variable>>;
+}
